Understanding the concept of loose and tight coupling
Using the need to implement a search feature that uses Binary search, we demonstrate how tight coupling is not good for cases where we may require a different sorting algorithm for our binary search feature. We then employ P2I and create flexibility regarding the use of various sorting algorithms in without changing the binary search code and with very minimal change in the sort parameter passed to the search call.